debug layer bug fixed. there was no uav barriers so it occured the bugs

+++ FullyGpuParticleSystem/Include/Util/GameTimer.h
@@ -7,7 +7,7 @@ public:
 	GameTimer();
 
 	float totalTime() const;
-	float deltaTime() const;
+	double deltaTime() const;
 
 	void reset();
 	void start();
